refactor transaction queue into a separate service
