+++ background/services/chain/tests/index.unit.test.ts
@@ -1,13 +1,12 @@
 import sinon from "sinon"
-import ChainService, { QueuedTxToRetrieve } from ".."
-import { ETHEREUM, MINUTE, OPTIMISM, POLYGON, SECOND } from "../../../constants"
+import ChainService from ".."
+import { ETHEREUM, MINUTE, OPTIMISM, POLYGON } from "../../../constants"
 import { EVMNetwork } from "../../../networks"
 import * as gas from "../../../lib/gas"
 import {
   createAddressOnNetwork,
   createBlockPrices,
   createChainService,
-  createTransactionsToRetrieve,
 } from "../../../tests/factories"
 import { UNIXTime } from "../../../types"
 import {
@@ -15,6 +14,7 @@ import {
   EnrichedLegacyTransactionSignatureRequest,
 } from "../../enrichment"
 import { AddressOnNetwork } from "../../../accounts"
+import { QueuedTxToRetrieve } from "../transaction-retrieve-queue.service"
 
 type ChainServiceExternalized = Omit<ChainService, ""> & {
   populatePartialEIP1559TransactionRequest: () => void
@@ -231,94 +231,6 @@ describe("Chain Service", () => {
       await chainService.getTrackedNetworks()
 
       expect(activeNetworksMock).toEqual([ETHEREUM, POLYGON])
-    })
-  })
-  describe("Queued Transaction Retrieve", () => {
-    describe("handleQueuedTransactionAlarm", () => {
-      let clock: sinon.SinonFakeTimers
-      let setIntervalSpy: sinon.SinonSpy
-      let chainServiceExternalized: ChainServiceExternalized
-      let retrieveTransactionStub: sinon.SinonStub
-      beforeEach(() => {
-        clock = sinon.useFakeTimers()
-
-        setIntervalSpy = sinon.spy(global, "setInterval")
-
-        chainServiceExternalized =
-          chainService as unknown as ChainServiceExternalized
-
-        retrieveTransactionStub = sandbox.stub(
-          chainServiceExternalized,
-          "retrieveTransaction"
-        )
-      })
-      afterEach(() => {
-        clock.restore()
-      })
-      it("should not start the granular timer if the queue is empty", () => {
-        chainServiceExternalized.handleQueuedTransactionAlarm()
-
-        clock.tick(2 * SECOND)
-
-        expect(setIntervalSpy.calledOnce).toBe(false)
-        expect(retrieveTransactionStub.callCount).toBe(0)
-        expect(chainServiceExternalized.transactionsToRetrieve.length).toBe(0)
-      })
-      it("should not recreate the timer when the alarm fires periodically", () => {
-        const clearIntervalSpy = sinon.spy(global, "clearInterval")
-
-        chainServiceExternalized.transactionsToRetrieve =
-          createTransactionsToRetrieve(100)
-
-        chainServiceExternalized.handleQueuedTransactionAlarm()
-        clock.tick(60 * SECOND)
-        chainServiceExternalized.handleQueuedTransactionAlarm()
-        clock.tick(60 * SECOND)
-        chainServiceExternalized.handleQueuedTransactionAlarm()
-        clock.tick(60 * SECOND)
-
-        expect(setIntervalSpy.calledOnce).toBe(true)
-        expect(clearIntervalSpy.calledOnce).toBe(false)
-      })
-      it("should retrieve 1 tx every 2 seconds, remove the tx from the queue and call the retrieve function", async () => {
-        const txInQueueCount = 100
-        const txRetrievedCount = 98
-
-        chainServiceExternalized.transactionsToRetrieve =
-          createTransactionsToRetrieve(txInQueueCount)
-
-        chainServiceExternalized.handleQueuedTransactionAlarm()
-
-        clock.tick(txRetrievedCount * 2 * SECOND)
-
-        expect(retrieveTransactionStub.callCount).toBe(txRetrievedCount)
-        expect(chainServiceExternalized.transactionsToRetrieve.length).toBe(
-          txInQueueCount - txRetrievedCount
-        )
-      })
-      it("should clean up the timer after the queue is emptied", async () => {
-        const clearIntervalSpy = sinon.spy(global, "clearInterval")
-        const numberOfTxInQueue = 100
-
-        chainServiceExternalized.transactionsToRetrieve =
-          createTransactionsToRetrieve(numberOfTxInQueue)
-
-        chainServiceExternalized.handleQueuedTransactionAlarm()
-
-        clock.tick(numberOfTxInQueue * 2 * SECOND)
-
-        expect(setIntervalSpy.calledOnce).toBe(true)
-        expect(retrieveTransactionStub.callCount).toBe(numberOfTxInQueue)
-        expect(chainServiceExternalized.transactionsToRetrieve.length).toBe(0)
-
-        // the clean up happens on the next tick
-        clock.tick(2 * SECOND)
-
-        expect(clearIntervalSpy.calledOnce).toBe(true)
-        expect(
-          chainServiceExternalized.transactionToRetrieveGranularTimer
-        ).toBe(undefined)
-      })
     })
   })
 })
